Update and rename README.md to Arduino_Coding_By_Rakshagan
Nowadays, attendance in schools and colleges mostly is based on paper. Sometimes this process causes errors and takes more time.

This project uses RFID technology to make a note of every student entering a classroom and also calculates the time resides in a class.

In this system, every student is allotted an RFID tag. The attendance can be done by placing the card near the RFID reader.